# SQLStore performance (avoid DISTINCT, filesort, add index), refs 2461