This PR is made in reference to: #2461, #2516, #2040

This PR addresses or contains:

- Avoid `SELECT DISTINCT ... ORDER BY ...` as it will most certainly cause a filesort and in worst case scenarios forces a full table scan [0]  which will __not__ be O(N) for large tables
- Add additional indices in order for the query planner to pick the best available option
- Not all queries will select the best query plan due to missing indices but current changes show a significant improvement over the existing setup
- `$GLOBALS['smwgQueryResultCacheType'] = 'redis';` was disabled during the test
- Some preliminary results include (`12802.0909ms` ↘ `10.7570ms`), (`22218.1082ms` ↘ `12.7010ms`) for ([content](https://gist.github.com/mwjames/103afa6ecd53b666de075e69cc1970ca) was used to simulate some high data load):

```
smw_object_ids: 169975 rows
smw_di_wikipage: 1404815 rows
smw_di_time: 140828 rows
smw_ft_search: 516884 rows
...
```

[0] https://dev.mysql.com/doc/refman/5.7/en/table-scan-avoidance.html

This PR includes:
- [ ] Tests (unit/integration)
- [ ] CI build passed