@mwjames 2017-06-17
Some unscientific benchmarks:

## Property page

Property: Has page with usage count 516310

BEFORE: (12802.0909ms)

```
SELECT DISTINCT smw_title, smw_namespace, smw_iw, smw_subobject, smw_sortkey, smw_sort FROM `smw_object_ids` INNER JOIN `smw_di_wikipage` AS t1 ON t1.s_id=smw_id WHERE t1.p_id='64' AND smw_iw!=':smw' AND smw_iw!=':smw-delete' ORDER BY smw_sort LIMIT 25	12802.0909ms	SMWSQLStore3Readers::getPropertySubjects
```

AFTER:  (10.7570ms)

```
SELECT smw_title, smw_namespace, smw_iw, smw_subobject, smw_sortkey, smw_sort FROM `smw_object_ids` INNER JOIN `smw_di_wikipage` AS t1 ON t1.s_id=smw_id WHERE t1.p_id='64' AND smw_iw!=':smw' AND smw_iw!=':smw-delete' ORDER BY smw_sort LIMIT 25	10.7570ms	SMWSQLStore3Readers::getPropertySubjects
```

## Query

```
smw_object_ids: 169975 rows
smw_di_wikipage: 1404815 rows
smw_di_time: 140828 rows
smw_ft_search: 516884 rows
```

BEFORE: (22218.1082ms)

```
SELECT DISTINCT t0.smw_id AS id,t0.smw_title AS t,t0.smw_namespace AS ns,t0.smw_iw AS iw,t0.smw_subobject AS so,t0.smw_sortkey AS sortkey,t5.o_sortkey FROM `smw_object_ids` AS t0 INNER JOIN `smw_di_wikipage` AS t1 ON t0.smw_id=t1.s_id INNER JOIN `smw_di_time` AS t5 ON t0.smw_id=t5.s_id WHERE (t1.p_id='64' AND t1.o_id='104') AND (t5.p_id='61') AND t0.smw_iw!=':smw' AND t0.smw_iw!=':smw-delete' AND t0.smw_iw!=':smw-redi' ORDER BY t5.o_sortkey ASC LIMIT 55	22218.1082ms	SMW\SQLStore\QueryEngine\QueryEngine::getInstanceQueryResult
```

AFTER: (12.7010ms)

```
SELECT t0.smw_id AS id,t0.smw_title AS t,t0.smw_namespace AS ns,t0.smw_iw AS iw,t0.smw_subobject AS so,t0.smw_sortkey AS sortkey,t5.o_sortkey FROM `smw_object_ids` AS t0 INNER JOIN `smw_di_wikipage` AS t1 ON t0.smw_id=t1.s_id INNER JOIN `smw_di_time` AS t5 ON t0.smw_id=t5.s_id WHERE (t1.p_id='64' AND t1.o_id='104') AND (t5.p_id='61') AND t0.smw_iw!=':smw' AND t0.smw_iw!=':smw-delete' AND t0.smw_iw!=':smw-redi' ORDER BY t5.o_sortkey ASC LIMIT 55	12.7010ms	SMW\SQLStore\QueryEngine\QueryEngine::getInstanceQueryResult
```