@mwjames 2017-06-17
In case `DISTINCT` isn't used you are required to compensate for duplicate matches and can only be dealt with using `GROUP BY` which isn't much better compared to `DISTINCT` as shown in:

14479.8319ms

```
SELECT t0.smw_id AS id,t0.smw_title AS t,t0.smw_namespace AS ns,t0.smw_iw AS iw,t0.smw_subobject AS so,t0.smw_sortkey AS sortkey,t5.o_sortkey FROM `smw_object_ids` AS t0 INNER JOIN `smw_di_wikipage` AS t1 ON t0.smw_id=t1.s_id INNER JOIN `smw_di_time` AS t5 ON t0.smw_id=t5.s_id WHERE (t1.p_id='64' AND t1.o_id='104') AND (t5.p_id='61') AND t0.smw_iw!=':smw' AND t0.smw_iw!=':smw-delete' AND t0.smw_iw!=':smw-redi' GROUP BY t0.smw_id ORDER BY t5.o_sortkey ASC LIMIT 55	14479.8319ms	SMW\SQLStore\QueryEngine\QueryEngine::getInstanceQueryResult
```

Sorting these matches is really killing the party here, the only other way would be to denormalize the tables and add the `smw_sort` to each one to avoid sorting on `smw_object_ids` with a full table scan.