@mwjames 2018-07-14
Closing this for now as it needs more investigation.